Fixes to free object list management to prevent objects getting overwritten. Removed the object in use flag. Wands now maintain their links to spell objects once more again until the next time it breaks

+++ Assets/scripts/Events/event_set_goal.cs
@@ -27,7 +27,7 @@
         {
             for (int i = 0; i <= npcs.GetUpperBound(0); i++)
             {
-                npcs[i].npc_goal = (short)newGoal;
+                npcs[i].npc_goal = (byte)newGoal;
                 npcs[i].npc_gtarg = (short)gtarg;
             }
         }
