+++ Assets/scripts/Magic/SpellEffects/SpellEffectAlly.cs
@@ -33,7 +33,7 @@ public class SpellEffectAlly : SpellEffect
                 //npc.state=NPC.AI_STATE_STANDING;	//Temporarily just wander around
                 npc.npc_attitude = NPC.AI_ATTITUDE_MELLOW;
                 //Makes the NPC follow the player
-                npc.npc_goal = (short)NPC.npc_goals.npc_goal_follow; ;
+                npc.npc_goal = (byte)NPC.npc_goals.npc_goal_follow; ;
                 npc.npc_gtarg = 1;
 
                 WasActive = true;
@@ -48,7 +48,7 @@ public class SpellEffectAlly : SpellEffect
         {
             //npc.state=OriginalState;
             npc.npc_attitude = OriginalAttitude;
-            npc.npc_goal = OriginalGoal;
+            npc.npc_goal = (byte)OriginalGoal;
             npc.npc_gtarg = OriginalGtarg;
         }
         base.CancelEffect();
